requirements.txt
# altair==5.5.0
# attrs==25.3.0
# blinker==1.9.0
# cachetools==5.5.2
# certifi==2025.1.31
# charset-normalizer==3.4.1
# click==8.1.8
# contourpy==1.3.0
# cycler==0.12.1
# filetype==1.2.0
# fonttools==4.56.0
# fpdf==1.7.2
# gitdb==4.0.12
# GitPython==3.1.44
# idna==3.7
# importlib_resources==6.5.2
# Jinja2==3.1.6
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# kiwisolver==1.4.7
# MarkupSafe==3.0.2
# matplotlib==3.9.4
# narwhals==1.32.0
# numpy==2.0.2
# opencv-python==4.11.0.86
# opencv-python-headless==4.10.0.84
# packaging==24.2
# pandas==2.2.3
# pillow==11.1.0
# pillow_heif==0.22.0
# protobuf==5.29.4
# pyarrow==19.0.1
# pydeck==0.9.1
# pyparsing==3.2.3
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# pytz==2025.2
# PyYAML==6.0.2
# referencing==0.36.2
# requests==2.32.3
# requests-toolbelt==1.0.0
# roboflow==1.1.58
# rpds-py==0.24.0
# six==1.17.0
# smmap==5.0.2
# streamlit==1.44.0
# tenacity==9.0.0
# toml==0.10.2
# tornado==6.4.2
# tqdm==4.67.1
# typing_extensions==4.13.0
# tzdata==2025.2
# urllib3==2.3.0
# zipp==3.21.0
streamlit
opencv-python-headless
pillow
fpdf
numpy
inference-sdk